Az bot: add BotPublishHelper, rename BotConnectionsManager to ServiceProviderManager
